# speed up the lab tests api load with select related and fewer queries