@davidmiller 2020-08-26
I'd imagine indexing the name would have a larger impact, but would need testing obv.